Cleaning up code based on comments
