+++ lib/cbdocument.utils.js
@@ -1,5 +1,6 @@
 import _ from 'lodash';
 import mpath from 'mpath';
+import { expandValues, processValue } from './utils';
 
 import Driver from 'couchbase-driver';
 
@@ -132,70 +133,15 @@ exports.buildRefValues = function (indexes, data) {
       const name = indexes[v].name;
       const indexType = indexes[v].indexType;
 
-      const processValue = initialValue => {
-        let finalVal = null;
-        if (typeof initialValue === 'string') {
-          finalVal = initialValue;
-        }
-        if (typeof initialValue === 'number') {
-          finalVal = initialValue.toString();
-        }
-        if (typeof initialValue === 'object' && initialValue._isBaseModel) {
-          finalVal = initialValue.getDocumentKeyValue(initialValue.config.storeFullReferenceId);
-        } else if (Array.isArray(initialValue)) {
-          finalVal = _.map(initialValue, v => {
-            if (v) {
-              if (typeof v === 'string') {
-                return v;
-              }
-              if (typeof v === 'number') {
-                return v.toString();
-              }
-              if (typeof v === 'object' && v._isBaseModel) {
-                return v.getDocumentKeyValue(v.config.storeFullReferenceId);
-              }
-            }
-          });
-
-          finalVal = _.compact(finalVal);
-          finalVal = finalVal.sort();
-        } else if (_.isUndefined(initialValue) || _.isNull(initialValue)) {
-          finalVal = null;
-        }
-        return finalVal;
-      }
-
       let value = null;
       if (indexes[v].compound) {
-        value = _.map(path, p => {
-          return processValue(mpath.get(p, data));
-        });
-
-        const expandValues = valueList => {
-          var retVal = [];
-          _.each(valueList, val => {
-            if (Array.isArray(val)) {
-              let expVal = expandValues(val);
-              let intRet = [];
-              _.each(expVal, v => {
-                intRet.push(retVal.concat(v));
-              });
-              retVal = intRet;
-            } else {
-              retVal.push(val);
-            }
-          });
-          return retVal;
-        }
-
+        value = _.map(path, p => processValue(mpath.get(p, data)));
         value = expandValues(value);
         value = _.compact(value);
 
-        let containsNested = _.some(value, _.isArray);
+        const containsNested = _.some(value, _.isArray);
         if (containsNested) {
-          value = _.map(value, v => {
-            return v.join('_');
-          });
+          value = _.map(value, v => v.join('_'));
         } else {
           value = value.join('_');
         }
@@ -206,7 +152,7 @@ exports.buildRefValues = function (indexes, data) {
 
       ret[name] = {
         path,
-        value: value,
+        value,
         name,
         indexType
       };
