+++ lib/schema.js
@@ -6,6 +6,8 @@ import * as utils from './utils';
 import * as schemaUtils from './schema.utils';
 import { normalizeProperties } from './normalize';
 
+const trailingDigitReg = new RegExp('^\\d+$');
+
 export default class Schema {
   /**
    * @classdesc Schema class represents the schema definition. It includes properties, methods, static methods, and any
@@ -198,9 +200,8 @@ export default class Schema {
       let indexType = options.indexType ? options.indexType: 'single';
       let indexName = options.indexName;
       if (!indexName){
-        const reg = new RegExp('^\\d+$');
         indexName = _.map(prop, p => {
-          if (reg.test(p)) {
+          if (trailingDigitReg.test(p)) {
             name = p[p.length - 2];
           }
           return inflection.singularize(p);
