+++ lib/utils.js
@@ -319,3 +319,53 @@ export function promisifyCall(ctx, fn, ...args) {
     fn.apply(ctx, args);
   });
 }
+
+export function expandValues(valueList) {
+  let retVal = [];
+  _.each(valueList, val => {
+    if (Array.isArray(val)) {
+      let expVal = expandValues(val);
+      let intRet = [];
+      _.each(expVal, v => {
+        intRet.push(retVal.concat(v));
+      });
+      retVal = intRet;
+    } else {
+      retVal.push(val);
+    }
+  });
+  return retVal;
+}
+
+export function processValue(initialValue) {
+  let finalVal = null;
+  if (typeof initialValue === 'string') {
+    finalVal = initialValue;
+  }
+  if (typeof initialValue === 'number') {
+    finalVal = initialValue.toString();
+  }
+  if (typeof initialValue === 'object' && initialValue._isBaseModel) {
+    finalVal = initialValue.getDocumentKeyValue(initialValue.config.storeFullReferenceId);
+  } else if (Array.isArray(initialValue)) {
+    finalVal = _.map(initialValue, v => {
+      if (v) {
+        if (typeof v === 'string') {
+          return v;
+        }
+        if (typeof v === 'number') {
+          return v.toString();
+        }
+        if (typeof v === 'object' && v._isBaseModel) {
+          return v.getDocumentKeyValue(v.config.storeFullReferenceId);
+        }
+      }
+    });
+
+    finalVal = _.compact(finalVal);
+    finalVal = finalVal.sort();
+  } else if (_.isUndefined(initialValue) || _.isNull(initialValue)) {
+    finalVal = null;
+  }
+  return finalVal;
+}
